# Optimize image conversion for half and float formats.