Optimizes conversion between RGBA variants of half and float image formats.
On average, this makes the conversion process ~7 times faster, depending on the formats. This is especially important for RGB images, which need to be converted to RGBA by most GPUs.

Build - Windows 64-bit, production
CPU: Ryzen 9 5900X 12-Core
RAM: 64 GB DDR4 3000 MHz

4096x4096 image:
| Conversion | Unopt | Opt |
|-------------|--------|-----|
| RH - RGH | 133ms | 14ms |
| RH - RGBH | 157ms | 28ms |
| RH - RGBAH | 176ms | 31ms |
| RGBH - RGBAH | 226ms | 32ms |
| RGBAH - RH | 163ms | 10ms |
| RGBAH - RGBH | 214ms| 22ms |

MRP: [image-half-convert.zip](https://github.com/godotengine/godot/files/15431381/image-half-convert.zip)