@BlueCube3310 2024-06-13
> Tested locally, it works as expected.

Thank you for the review!

> I'm a bit surprised I don't see any performance improvement in this scenario, but at least it's not slower than before.

The first conversion happens between the source 8-bit uint image and the specified half/float format. In order to get the correct results, only the second conversion/match statement has to be measured.
This PR only affects the conversion times between halfX-halfY and floatX-floatY conversions, since these ones have the same data type and can be handled with a simple memcpy, instead of the slower fallback function.